refactor: operator 패키지 위치 model 하위로 변경
